fix: Resolve critical crash in project loading and refactor data model

+++ src/core/graph/Connection.h
@@ -2,17 +2,18 @@
 #define CONNECTION_H
 #include <string>
 #include <utility>
+#include <cstdint>
 
 struct Connection {
     double rate = 0.0;
-    int id;
-    int from_port; // Port number on the originating node
-    int to_port; // Port number on the destination node
+    uint64_t id;
+    uint64_t from_port; // Port number on the originating node
+    uint64_t to_port; // Port number on the destination node
     std::string resource_key;
 
     Connection() = default;
 
-    Connection(int id,int from_port, int to_port, std::string resource_key)
+    Connection(uint64_t id, uint64_t from_port, uint64_t to_port, std::string resource_key)
         : id(id), from_port(from_port), to_port(to_port), resource_key(std::move(resource_key)) {
     }
 
@@ -29,7 +30,7 @@ namespace std {
     template <>
     struct hash<Connection> {
         std::size_t operator()(const Connection& c) const noexcept {
-            return std::hash<int>{}(c.id);
+            return std::hash<uint64_t>{}(c.id);
         }
     };
 }
